requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
google-genai==1.11.0
python-dotenv==1.0.1
